usb-device-manager: Fix number truncation by spice_usb_device_get_busnum
"bus" field is 16 bit so returning using a guint8 type potentially
truncate the value.

Acked-by: Victor Toso <victortoso@redhat.com>

+++ src/usb-device-manager-priv.h
@@ -29,7 +29,7 @@ G_BEGIN_DECLS
 void spice_usb_device_manager_device_error(
     SpiceUsbDeviceManager *manager, SpiceUsbDevice *device, GError *err);
 
-guint8 spice_usb_device_get_busnum(const SpiceUsbDevice *device);
+guint16 spice_usb_device_get_busnum(const SpiceUsbDevice *device);
 guint8 spice_usb_device_get_devaddr(const SpiceUsbDevice *device);
 guint16 spice_usb_device_get_vid(const SpiceUsbDevice *device);
 guint16 spice_usb_device_get_pid(const SpiceUsbDevice *device);
